backend/pyproject.toml
[project]
name = "agent-eval-backend"
version = "0.1.0"
description = "Backend service for the agent stability evaluation tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.110.1",
    "uvicorn[standard]==0.29.0",
    "pydantic-settings==2.2.1",
    "SQLAlchemy==2.0.29",
    "alembic==1.13.1",
    "psycopg2-binary==2.9.9",
    "celery==5.3.6",
    "redis==5.0.3",
    "httpx==0.27.0",
    "pandas>=2.3.0,<2.4",
    "openpyxl==3.1.2",
    "xlrd==2.0.1",
    "python-multipart==0.0.9",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "tenacity==8.2.3"
]

[project.optional-dependencies]
dev = [
    "pytest==8.1.1",
    "pytest-asyncio==0.23.5",
    "pytest-cov==4.1.0",
    "ruff==0.3.5",
    "mypy==1.9.0",
    "factory-boy==3.3.0"
]

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
